requirements.txt
fastapi
pydantic-settings
sqlalchemy
uvicorn
pyjwt
bcrypt
pytz
ujson